+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.27.2 - Unreleased
+
 ## 0.27.1 - 2026-06-15
 
 ### Fixed